requirements.txt
psycopg2-binary
beautifulsoup4
pandas
polars
openpyxl
requests
